Move service_data to new files
